#132: Move Repository impl in ui module
