--- importer/__init__.py
@@ -1,1 +1,0 @@
-from .Repository import Repository, RepositoryTreeObject
