+++ ui/__init__.py
@@ -1,2 +1,6 @@
-from .TreeObject import TreeObject
 from .DirectoryObject import DirectoryObject
+from .RepositoryId import RepositoryId
+from .RepositoryTreeObject import RepositoryTreeObject
+from .RepositoryTree import RepositoryTree
+from .Repository import Repository
+from .TreeObject import TreeObject
